feat(dapp-kit): add README and refactor hooks structure
