+++ packages/dapp-kit/src/hooks/index.ts
@@ -1,16 +1,4 @@
-export * from './useHubProvider';
-export * from './useSpokeProvider';
-export * from './useWalletProvider';
-export * from './useSuppliedAssets';
-export * from './useHubWalletAddress';
-export * from './useSupply';
-export * from './useWithdraw';
-export * from './useSodaxContext';
-export * from './useQuote';
-export * from './useCreateIntentOrder';
-export * from './useStatus';
-export * from './useBorrow';
-export * from './useRepay';
-export * from './useReservesData';
-export * from './useAllowance';
-export * from './useApprove';
+export * from './shared';
+export * from './provider';
+export * from './mm';
+export * from './swap';
