+++ packages/dapp-kit/src/hooks/mm/useWithdraw.ts
@@ -9,23 +9,24 @@ import {
 import type { XChainId, XToken } from '@sodax/xwagmi';
 import { getXChainType, useXAccount, xChainMap } from '@sodax/xwagmi';
 import { useState } from 'react';
-import type { Address, Hex } from 'viem';
+import type { Address, Hash, Hex } from 'viem';
 import { parseUnits } from 'viem';
-import { useHubProvider } from './useHubProvider';
-import { useHubWalletAddress } from './useHubWalletAddress';
-import { useSpokeProvider } from './useSpokeProvider';
-import { useSodaxContext } from './useSodaxContext';
+import { useHubProvider } from '../provider/useHubProvider';
+import { useHubWalletAddress } from '../mm/useHubWalletAddress';
+import { useSpokeProvider } from '../provider/useSpokeProvider';
+import { useSodaxContext } from '../shared/useSodaxContext';
 import { XCALL_RELAY_URL } from '@/constants';
+import { getSpokeTokenAddressByVault } from '@/core';
 
-interface UseRepayReturn {
-  repay: (amount: string) => Promise<void>;
+interface UseWithdrawReturn {
+  withdraw: (amount: string) => Promise<void>;
   isLoading: boolean;
   error: Error | null;
   resetError: () => void;
 }
 
 // token: this is hub token
-export function useRepay(token: XToken, spokeChainId: XChainId): UseRepayReturn {
+export function useWithdraw(token: XToken, spokeChainId: XChainId): UseWithdrawReturn {
   const { address } = useXAccount(getXChainType(token.xChainId));
   const { sodax } = useSodaxContext();
   const hubProvider = useHubProvider();
@@ -40,7 +41,7 @@ export function useRepay(token: XToken, spokeChainId: XChainId): UseRepayReturn 
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [error, setError] = useState<Error | null>(null);
 
-  const repay = async (amount: string): Promise<void> => {
+  const withdraw = async (amount: string): Promise<void> => {
     if (!hubWalletAddress) {
       setError(new Error('hubWalletAddress is not found'));
       return;
@@ -58,20 +59,17 @@ export function useRepay(token: XToken, spokeChainId: XChainId): UseRepayReturn 
     setError(null);
 
     try {
-      const data: Hex = sodax.moneyMarket.repayData(
-        token.address,
+      const data: Hex = sodax.moneyMarket.withdrawData(
         hubWalletAddress as Address,
+        spokeProvider.walletProvider.getWalletAddress(),
+        getSpokeTokenAddressByVault(spokeChainId, token.address),
         parseUnits(amount, token.decimals),
         spokeProvider.chainConfig.chain.id,
       );
 
-      const txHash = await SpokeService.deposit(
-        {
-          from: address as `0x${string}`,
-          token: token.address as `0x${string}`,
-          amount: parseUnits(amount, token.decimals),
-          data,
-        },
+      const txHash: Hash = await SpokeService.callWallet(
+        spokeProvider.walletProvider.getWalletAddress(),
+        data,
         spokeProvider,
         hubProvider,
       );
@@ -103,7 +101,7 @@ export function useRepay(token: XToken, spokeChainId: XChainId): UseRepayReturn 
   };
 
   return {
-    repay,
+    withdraw,
     isLoading,
     error,
     resetError,
