+++ packages/dapp-kit/src/hooks/shared/useApprove.ts
@@ -1,6 +1,6 @@
 import type { XToken } from '@sodax/xwagmi';
-import { useSodaxContext } from './useSodaxContext';
-import { useSpokeProvider } from './useSpokeProvider';
+import { useSodaxContext } from '../shared/useSodaxContext';
+import { useSpokeProvider } from '../provider/useSpokeProvider';
 import type { Address, SpokeChainId } from '@sodax/sdk';
 import { parseUnits } from 'viem';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
